Remove test_numerical.py
