+++ python/cudf/cudf/tests/general_functions/test_to_numeric.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2021-2024, NVIDIA CORPORATION.
+# Copyright (c) 2025, NVIDIA CORPORATION.
 
 import numpy as np
 import pandas as pd
@@ -7,122 +7,12 @@ import pytest
 import cudf
 from cudf.core._compat import PANDAS_GE_220
 from cudf.testing import assert_eq
-from cudf.testing._utils import NUMERIC_TYPES, expect_warning_if
-from cudf.utils.dtypes import np_dtypes_to_pandas_dtypes
+from cudf.testing._utils import expect_warning_if
 
 
-def test_can_cast_safely_same_kind():
-    # 'i' -> 'i'
-    data = cudf.Series([1, 2, 3], dtype="int32")._column
-    to_dtype = np.dtype("int64")
-
-    assert data.can_cast_safely(to_dtype)
-
-    data = cudf.Series([1, 2, 3], dtype="int64")._column
-    to_dtype = np.dtype("int32")
-
-    assert data.can_cast_safely(to_dtype)
-
-    data = cudf.Series([1, 2, 2**31], dtype="int64")._column
-    assert not data.can_cast_safely(to_dtype)
-
-    # 'u' -> 'u'
-    data = cudf.Series([1, 2, 3], dtype="uint32")._column
-    to_dtype = np.dtype("uint64")
-
-    assert data.can_cast_safely(to_dtype)
-
-    data = cudf.Series([1, 2, 3], dtype="uint64")._column
-    to_dtype = np.dtype("uint32")
-
-    assert data.can_cast_safely(to_dtype)
-
-    data = cudf.Series([1, 2, 2**33], dtype="uint64")._column
-    assert not data.can_cast_safely(to_dtype)
-
-    # 'f' -> 'f'
-    data = cudf.Series([np.inf, 1.0], dtype="float64")._column
-    to_dtype = np.dtype("float32")
-    assert data.can_cast_safely(to_dtype)
-
-    data = cudf.Series(
-        [float(np.finfo("float32").max) * 2, 1.0], dtype="float64"
-    )._column
-    to_dtype = np.dtype("float32")
-    assert not data.can_cast_safely(to_dtype)
-
-
-def test_can_cast_safely_mixed_kind():
-    data = cudf.Series([1, 2, 3], dtype="int32")._column
-    to_dtype = np.dtype("float32")
-    assert data.can_cast_safely(to_dtype)
-
-    # too big to fit into f32 exactly
-    data = cudf.Series([1, 2, 2**24 + 1], dtype="int32")._column
-    assert not data.can_cast_safely(to_dtype)
-
-    data = cudf.Series([1, 2, 3], dtype="uint32")._column
-    to_dtype = np.dtype("float32")
-    assert data.can_cast_safely(to_dtype)
-
-    # too big to fit into f32 exactly
-    data = cudf.Series([1, 2, 2**24 + 1], dtype="uint32")._column
-    assert not data.can_cast_safely(to_dtype)
-
-    to_dtype = np.dtype("float64")
-    assert data.can_cast_safely(to_dtype)
-
-    data = cudf.Series([1.0, 2.0, 3.0], dtype="float32")._column
-    to_dtype = np.dtype("int32")
-    assert data.can_cast_safely(to_dtype)
-
-    # not integer float
-    data = cudf.Series([1.0, 2.0, 3.5], dtype="float32")._column
-    assert not data.can_cast_safely(to_dtype)
-
-    data = cudf.Series([10.0, 11.0, 2000.0], dtype="float64")._column
-    assert data.can_cast_safely(to_dtype)
-
-    # float out of int range
-    data = cudf.Series([1.0, 2.0, 1.0 * (2**31)], dtype="float32")._column
-    assert not data.can_cast_safely(to_dtype)
-
-    # negative signed integers casting to unsigned integers
-    data = cudf.Series([-1, 0, 1], dtype="int32")._column
-    to_dtype = np.dtype("uint32")
-    assert not data.can_cast_safely(to_dtype)
-
-
-def test_to_pandas_nullable_integer():
-    gsr_not_null = cudf.Series([1, 2, 3])
-    gsr_has_null = cudf.Series([1, 2, None])
-
-    psr_not_null = pd.Series([1, 2, 3], dtype="int64")
-    psr_has_null = pd.Series([1, 2, None], dtype="Int64")
-
-    assert_eq(gsr_not_null.to_pandas(), psr_not_null)
-    assert_eq(gsr_has_null.to_pandas(nullable=True), psr_has_null)
-
-
-def test_to_pandas_nullable_bool():
-    gsr_not_null = cudf.Series([True, False, True])
-    gsr_has_null = cudf.Series([True, False, None])
-
-    psr_not_null = pd.Series([True, False, True], dtype="bool")
-    psr_has_null = pd.Series([True, False, None], dtype="boolean")
-
-    assert_eq(gsr_not_null.to_pandas(), psr_not_null)
-    assert_eq(gsr_has_null.to_pandas(nullable=True), psr_has_null)
-
-
-def test_can_cast_safely_has_nulls():
-    data = cudf.Series([1, 2, 3, None], dtype="float32")._column
-    to_dtype = np.dtype("int64")
-
-    assert data.can_cast_safely(to_dtype)
-
-    data = cudf.Series([1, 2, 3.1, None], dtype="float32")._column
-    assert not data.can_cast_safely(to_dtype)
+@pytest.fixture(params=["integer", "signed", "unsigned", "float"])
+def downcast(request):
+    return request.param
 
 
 @pytest.mark.parametrize(
@@ -182,9 +72,6 @@ def test_to_numeric_basic_1d(data):
         [np.iinfo(np.int64).max, np.iinfo(np.int64).min],
     ],
 )
-@pytest.mark.parametrize(
-    "downcast", ["integer", "signed", "unsigned", "float"]
-)
 def test_to_numeric_downcast_int(data, downcast):
     ps = pd.Series(data)
     gs = cudf.from_pandas(ps)
@@ -212,9 +99,6 @@ def test_to_numeric_downcast_int(data, downcast):
         [1.0, 1.5, 2.6, 3.4],
     ],
 )
-@pytest.mark.parametrize(
-    "downcast", ["signed", "integer", "unsigned", "float"]
-)
 def test_to_numeric_downcast_float(data, downcast):
     ps = pd.Series(data)
     gs = cudf.from_pandas(ps)
@@ -237,8 +121,10 @@ def test_to_numeric_downcast_float(data, downcast):
         [-1.0, -1.79e308],
     ],
 )
-@pytest.mark.parametrize("downcast", ["signed", "integer", "unsigned"])
 def test_to_numeric_downcast_large_float(data, downcast):
+    if downcast == "float":
+        pytest.skip(f"{downcast=} not applicable for test")
+
     ps = pd.Series(data)
     gs = cudf.from_pandas(ps)
 
@@ -260,8 +146,8 @@ def test_to_numeric_downcast_large_float(data, downcast):
         [-1.0, -1.79e308],
     ],
 )
-@pytest.mark.parametrize("downcast", ["float"])
-def test_to_numeric_downcast_large_float_pd_bug(data, downcast):
+def test_to_numeric_downcast_large_float_pd_bug(data):
+    downcast = "float"
     ps = pd.Series(data)
     gs = cudf.from_pandas(ps)
 
@@ -277,9 +163,6 @@ def test_to_numeric_downcast_large_float_pd_bug(data, downcast):
         ["1", "2", "3"],
         [str(np.iinfo(np.int64).max), str(np.iinfo(np.int64).min)],
     ],
-)
-@pytest.mark.parametrize(
-    "downcast", ["signed", "integer", "unsigned", "float"]
 )
 def test_to_numeric_downcast_string_int(data, downcast):
     ps = pd.Series(data)
@@ -301,9 +184,6 @@ def test_to_numeric_downcast_string_int(data, downcast):
         ["1", "10", "1.0", "2e3", "2e+3", "2e-3"],
         ["1", "10", "1.0", "2e3", "", ""],  # mixed empty strings
     ],
-)
-@pytest.mark.parametrize(
-    "downcast", ["signed", "integer", "unsigned", "float"]
 )
 def test_to_numeric_downcast_string_float(data, downcast):
     ps = pd.Series(data)
@@ -334,9 +214,6 @@ def test_to_numeric_downcast_string_float(data, downcast):
             "-1.79769313486231e308",
         ],  # 2 digits relaxed from np.finfo(np.float64).min/max
     ],
-)
-@pytest.mark.parametrize(
-    "downcast", ["signed", "integer", "unsigned", "float"]
 )
 def test_to_numeric_downcast_string_large_float(data, downcast):
     ps = pd.Series(data)
@@ -382,35 +259,8 @@ def test_to_numeric_error(data, errors):
         assert_eq(expect, got)
 
 
-@pytest.mark.parametrize("dtype", NUMERIC_TYPES)
-@pytest.mark.parametrize("input_obj", [[1, cudf.NA, 3]])
-def test_series_construction_with_nulls(dtype, input_obj):
-    dtype = cudf.dtype(dtype)
-    # numpy case
-
-    expect = pd.Series(input_obj, dtype=np_dtypes_to_pandas_dtypes[dtype])
-    got = cudf.Series(input_obj, dtype=dtype).to_pandas(nullable=True)
-
-    assert_eq(expect, got)
-
-    # Test numpy array of objects case
-    np_data = [
-        dtype.type(v) if v is not cudf.NA else cudf.NA for v in input_obj
-    ]
-
-    expect = pd.Series(np_data, dtype=np_dtypes_to_pandas_dtypes[dtype])
-    got = cudf.Series(np_data, dtype=dtype).to_pandas(nullable=True)
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [[True, False, True]],
-)
-@pytest.mark.parametrize(
-    "downcast", ["signed", "integer", "unsigned", "float"]
-)
-def test_series_to_numeric_bool(data, downcast):
+def test_series_to_numeric_bool(downcast):
+    data = [True, False, True]
     ps = pd.Series(data)
     gs = cudf.from_pandas(ps)
