+++ python/cudf/cudf/tests/series/methods/test_to_pandas.py
@@ -135,6 +135,28 @@ def test_series_to_pandas_arrow_type_nullable_raises(scalar):
         ser.to_pandas(nullable=True, arrow_type=True)
 
 
+def test_to_pandas_nullable_integer():
+    gsr_not_null = cudf.Series([1, 2, 3])
+    gsr_has_null = cudf.Series([1, 2, None])
+
+    psr_not_null = pd.Series([1, 2, 3], dtype="int64")
+    psr_has_null = pd.Series([1, 2, None], dtype="Int64")
+
+    assert_eq(gsr_not_null.to_pandas(), psr_not_null)
+    assert_eq(gsr_has_null.to_pandas(nullable=True), psr_has_null)
+
+
+def test_to_pandas_nullable_bool():
+    gsr_not_null = cudf.Series([True, False, True])
+    gsr_has_null = cudf.Series([True, False, None])
+
+    psr_not_null = pd.Series([True, False, True], dtype="bool")
+    psr_has_null = pd.Series([True, False, None], dtype="boolean")
+
+    assert_eq(gsr_not_null.to_pandas(), psr_not_null)
+    assert_eq(gsr_has_null.to_pandas(nullable=True), psr_has_null)
+
+
 @pytest.mark.parametrize(
     "scalar",
     [
