+++ python/cudf/cudf/tests/series/test_constructors.py
@@ -21,6 +21,7 @@ from cudf.core.column.column import as_column
 from cudf.errors import MixedTypeError
 from cudf.testing import assert_eq
 from cudf.testing._utils import assert_exceptions_equal
+from cudf.utils.dtypes import np_dtypes_to_pandas_dtypes
 
 
 @pytest.mark.parametrize(
@@ -631,6 +632,26 @@ def test_construct_nonnative_array(arr):
     assert_eq(result, expected)
 
 
+@pytest.mark.parametrize("input_obj", [[1, cudf.NA, 3]])
+def test_series_construction_with_nulls(numeric_types_as_str, input_obj):
+    dtype = np.dtype(numeric_types_as_str)
+    # numpy case
+
+    expect = pd.Series(input_obj, dtype=np_dtypes_to_pandas_dtypes[dtype])
+    got = cudf.Series(input_obj, dtype=dtype).to_pandas(nullable=True)
+
+    assert_eq(expect, got)
+
+    # Test numpy array of objects case
+    np_data = [
+        dtype.type(v) if v is not cudf.NA else cudf.NA for v in input_obj
+    ]
+
+    expect = pd.Series(np_data, dtype=np_dtypes_to_pandas_dtypes[dtype])
+    got = cudf.Series(np_data, dtype=dtype).to_pandas(nullable=True)
+    assert_eq(expect, got)
+
+
 @pytest.mark.parametrize("nan_as_null", [True, False])
 def test_construct_all_pd_NA_with_dtype(nan_as_null):
     result = cudf.Series(
@@ -1168,7 +1189,7 @@ def test_roundtrip_series_plc_column(ps):
     assert_eq(expect, actual)
 
 
-def test_series_construction_with_nulls():
+def test_series_structarray_construction_with_nulls():
     fields = [
         pa.array([1], type=pa.int64()),
         pa.array([None], type=pa.int64()),
